# [Core] Remove host GPU sync in `merge_multimodal_embeddings`